feat : rental create apply CLEAR (#S09P12D102-177)

+++ phonesin/src/main/java/com/regalaxy/phonesin/phone/model/entity/Model.java
@@ -15,7 +15,7 @@ public class Model {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "model_id")
-    private int model_id;
+    private Long model_id;
 
     private String model_name;
     private String nickname;
